requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Flask-Mail==0.9.1
Flask-Babel==2.0.0
Flask-Limiter==3.3.1
Flask-Security-Too==4.1.5
Werkzeug==2.2.31.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Migrate==4.1.0
Flask-Mail==0.10.0
Flask-Babel==3.1.0
Flask-Limiter==3.12.0
Flask-Security-Too==5.4.0
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
Werkzeug==3.0.1
twilio==9.6.2
pandas==2.2.1
reportlab==4.1.0
APScheduler==3.11.0
requests==2.31.0
python-dotenv==1.0.1
email_validator==2.2.0
cryptography==42.0.5
bcrypt==4.1.2
pgAudit==1.7.0
jsonschema==4.21.1
phonenumbers==8.13.30
Pillow==10.2.0
argon2-cffi==23.1.0
zxcvbn==4.4.28